refactor: hue, sat
